refactor(autonomi): move `send` to lib.rs
