+++ libautonomi/src/lib.rs
@@ -40,7 +40,7 @@ mod client {
         },
         NetworkAddress,
     };
-    use sn_transfers::Payment;
+    use sn_transfers::{Payment, SpendReason, Transfer};
     use tokio::{
         sync::mpsc::Receiver,
         task::{JoinError, JoinSet},
@@ -48,6 +48,7 @@ mod client {
     };
     use xor_name::XorName;
 
+    use crate::wallet::MemWallet;
     use crate::{
         client_wallet::SendSpendsError,
         self_encryption::{encrypt, DataMapLevel},
@@ -144,6 +145,44 @@ mod client {
             receiver.await.expect("sender should not close")?;
 
             Ok(Self { network })
+        }
+
+        /// Creates a `Transfer` that can be received by the receiver.
+        /// Once received, it will be turned into a `CashNote` that the receiver can spend.
+        pub async fn send(
+            &mut self,
+            to: MainPubkey,
+            amount_in_nano: NanoTokens,
+            reason: Option<SpendReason>,
+            wallet: &mut MemWallet,
+        ) -> eyre::Result<Transfer> {
+            let offline_transfer =
+                wallet.create_offline_transfer(vec![(amount_in_nano, to)], reason)?;
+
+            // return the first CashNote (assuming there is only one because we only sent to one recipient)
+            let cash_note_for_recipient = match &offline_transfer.cash_notes_for_recipient[..] {
+                [cash_note] => Ok(cash_note),
+                [_multiple, ..] => Err(SendSpendsError::CouldNotSendMoney(
+                    "Multiple CashNotes were returned from the transaction when only one was expected."
+                        .into(),
+                )),
+                [] => Err(SendSpendsError::CouldNotSendMoney(
+                    "No CashNotes were returned from the wallet.".into(),
+                )),
+            }?;
+
+            let transfer = Transfer::transfer_from_cash_note(cash_note_for_recipient)?;
+
+            self.send_spends(offline_transfer.all_spend_requests.iter())
+                .await?;
+
+            wallet.process_offline_transfer(offline_transfer.clone());
+
+            for spend in &offline_transfer.all_spend_requests {
+                wallet.add_pending_spend(spend.clone());
+            }
+
+            Ok(transfer)
         }
 
         async fn fetch_from_data_map(&self, data_map: &DataMap) -> Result<Bytes, GetError> {
